+++ changelog.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Bugs Fixed
 
+- Fixes Direct/TCP connection limit options to reject values below their documented minimums. See [PR 5883](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/5883).
+
 #### Other Changes
 
 ### <a name="3.61.0-preview.0"/> [3.61.0-preview.0](https://www.nuget.org/packages/Microsoft.Azure.Cosmos/3.61.0-preview.0) - 2026-5-18